Fix FRAME_DETAIL_MAPPER for Max RSS

+++ cuebot/src/main/java/com/imageworks/spcue/dao/postgres/FrameDaoJdbc.java
@@ -400,7 +400,7 @@ public class FrameDaoJdbc extends JdbcDaoSupport  implements FrameDao {
             frame.jobId = rs.getString("pk_job");
             frame.layerId = rs.getString("pk_layer");
             frame.showId = rs.getString("pk_show");
-            frame.maxRss = rs.getInt("int_mem_max_used");
+            frame.maxRss = rs.getLong("int_mem_max_used");
             frame.name = rs.getString("str_name");
             frame.number = rs.getInt("int_number");
             frame.dispatchOrder = rs.getInt("int_dispatch_order");
